Refactor AccountEntity
